Move trim_dataframe from history to converter
