+++ freqtrade/optimize/hyperopt.py
@@ -23,7 +23,8 @@ from joblib import (Parallel, cpu_count, delayed, dump, load,
 from pandas import DataFrame
 
 from freqtrade import OperationalException
-from freqtrade.data.history import get_timerange, trim_dataframe
+from freqtrade.data.history import get_timerange
+from freqtrade.data.converter import trim_dataframe
 from freqtrade.misc import plural, round_dict
 from freqtrade.optimize.backtesting import Backtesting
 # Import IHyperOpt and IHyperOptLoss to allow unpickling classes from these modules
